ref: move client-only response types to own folder
