+++ src/public/modules/forms/helpers/csv-response-classes/ArrayAnswerResponse.class.ts
@@ -1,4 +1,4 @@
-import { ArrayResponse } from '../../../../../types/response'
+import { ArrayResponse } from '../../../../types/response'
 
 import { Response } from './Response.class'
