+++ src/public/modules/forms/helpers/csv-response-classes/SingleAnswerResponse.class.ts
@@ -1,4 +1,4 @@
-import { SingleResponse } from '../../../../../types/response'
+import { SingleResponse } from '../../../../types/response'
 
 import { Response } from './Response.class'
