+++ src/public/types/README.md
@@ -1,0 +1,1 @@
+Contains types used solely in the client.
