+++ src/types/response/index.ts
@@ -56,21 +56,3 @@ export interface IClientEncryptSubmission extends IClientSubmission {
   encryptedContent: string
   version: number
 }
-
-export type DisplayedResponseWithoutAnswer = {
-  _id: string
-  question: string
-  fieldType: string
-  isHeader?: boolean
-}
-
-export type ArrayResponse = DisplayedResponseWithoutAnswer & {
-  answerArray: string[]
-}
-
-export type NestedResponse = DisplayedResponseWithoutAnswer & {
-  answerArray: string[][]
-}
-export type SingleResponse = DisplayedResponseWithoutAnswer & {
-  answer: string
-}
